Move cljs-navigation into the repo
Because the wrapper was on the outside repo it made the navigation logic to be scattered across multiple files. Changes and extensions were more complex to be added. Cause this library is just a basic wrapper, best place is to keep it nearby the code, which will made the updates more easier and group code better.